refactor panel into dedicated folder
