+++ extensions/vscode/.vscodeignore
@@ -18,6 +18,7 @@ vsc-extension-quickstart.md
 # keep these
 !assets/tree-sitter-silverscript.wasm
 !queries/**
+!webviews/**
 !node_modules/web-tree-sitter/package.json
 !node_modules/web-tree-sitter/LICENSE
 !node_modules/web-tree-sitter/web-tree-sitter.cjs
