+++ extensions/vscode/webviews/quickLaunch/panel.css
@@ -1,0 +1,262 @@
+:root {
+  --bg: var(--vscode-editor-background);
+  --fg: var(--vscode-editor-foreground);
+  --input-bg: var(--vscode-input-background);
+  --input-fg: var(--vscode-input-foreground);
+  --input-border: var(--vscode-input-border, transparent);
+  --panel-bg: rgba(127, 127, 127, 0.03);
+  --panel-hover: var(--vscode-toolbar-hoverBackground, rgba(128, 128, 128, 0.1));
+  --btn: var(--vscode-button-background);
+  --btn-fg: var(--vscode-button-foreground);
+  --btn-hover: var(--vscode-button-hoverBackground);
+  --btn-secondary-bg: transparent;
+  --btn-secondary-fg: var(--fg);
+  --btn-secondary-hover: var(--panel-hover);
+  --focus: var(--vscode-focusBorder);
+  --muted: rgba(127, 127, 127, 0.75);
+  --sep: var(--vscode-widget-border, rgba(128, 128, 128, 0.25));
+  --badge: var(--vscode-badge-background);
+  --badge-fg: var(--vscode-badge-foreground);
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  padding: 16px;
+  color: var(--fg);
+  background: var(--bg);
+  font: 13px/1.45 var(--vscode-font-family, system-ui, sans-serif);
+}
+
+.header-row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px;
+  margin: 0 0 14px;
+}
+
+h1 {
+  flex: 1;
+  min-width: 0;
+  margin: 0;
+  font-size: 16px;
+  font-weight: 600;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.topbar {
+  display: flex;
+  align-items: center;
+  flex: none;
+}
+
+.topbar-actions {
+  display: flex;
+  gap: 6px;
+}
+
+section {
+  margin-bottom: 14px;
+}
+
+h2 {
+  margin: 0 0 8px;
+  font-size: 11px;
+  font-weight: 700;
+  letter-spacing: 0.06em;
+  text-transform: uppercase;
+  color: var(--muted);
+}
+
+label {
+  display: block;
+  margin: 10px 0 4px;
+  font-size: 12px;
+}
+
+.meta {
+  color: var(--muted);
+  font-size: 11px;
+  margin-left: 6px;
+}
+
+.badge {
+  margin-left: 6px;
+  padding: 1px 5px;
+  border-radius: 3px;
+  background: var(--badge);
+  color: var(--badge-fg);
+  font-size: 10px;
+  font-weight: 600;
+  vertical-align: middle;
+}
+
+input,
+select {
+  width: 100%;
+  min-height: 34px;
+  padding: 6px 10px;
+  border-radius: 4px;
+  border: 1px solid var(--input-border);
+  background: var(--input-bg);
+  color: var(--input-fg);
+  outline: none;
+  font: 13px var(--vscode-editor-font-family, monospace);
+}
+
+input:focus,
+select:focus {
+  border-color: var(--focus);
+}
+
+.empty {
+  margin: 0;
+  color: var(--muted);
+  font-style: italic;
+}
+
+.actions {
+  display: flex;
+  gap: 8px;
+  margin-top: 16px;
+}
+
+button {
+  flex: 1;
+  min-height: 36px;
+  padding: 8px 0;
+  border: 0;
+  border-radius: 4px;
+  background: var(--btn);
+  color: var(--btn-fg);
+  font-size: 13px;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+button:hover {
+  background: var(--btn-hover);
+}
+
+.secondary-button {
+  flex: none;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: auto;
+  min-width: 0;
+  min-height: 32px;
+  padding: 0 12px;
+  border: 1px solid var(--sep);
+  border-radius: 4px;
+  background: var(--btn-secondary-bg);
+  color: var(--btn-secondary-fg);
+  font-size: 11px;
+  font-weight: 600;
+  line-height: 1;
+}
+
+.secondary-button:hover {
+  background: var(--btn-secondary-hover);
+}
+
+.compact-button {
+  flex: none;
+}
+
+.field-row {
+  position: relative;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.field-row input {
+  flex: 1;
+}
+
+.field-row input.crypto-input {
+  cursor: pointer;
+}
+
+.field-row .field-action {
+  min-width: 64px;
+  padding: 0 10px;
+}
+
+.identity-dropdown {
+  position: absolute;
+  z-index: 100;
+  top: calc(100% + 4px);
+  left: 0;
+  right: 0;
+  padding: 4px 0;
+  border: 1px solid var(--sep);
+  border-radius: 6px;
+  background: var(--input-bg);
+  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.28);
+}
+
+.identity-choice {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  justify-content: space-between;
+  margin: 0 4px;
+  padding: 7px 10px;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.identity-choice:hover {
+  background: var(--btn);
+  color: var(--btn-fg);
+}
+
+.identity-choice-name {
+  display: block;
+  font-weight: 700;
+}
+
+.identity-choice-main {
+  flex: 1;
+  min-width: 0;
+}
+
+.identity-choice-value {
+  display: block;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: 11px;
+  opacity: 0.72;
+}
+
+.identity-choice-delete {
+  flex: none;
+  min-width: 18px;
+  min-height: 18px;
+  padding: 0;
+  border: 0;
+  border-radius: 3px;
+  background: transparent;
+  color: var(--muted);
+  font-size: 14px;
+  line-height: 1;
+}
+
+.identity-choice-delete:hover {
+  background: var(--panel-hover);
+  color: var(--vscode-errorForeground, var(--fg));
+}
+
+.identity-divider {
+  margin: 4px 0;
+  border-top: 1px solid var(--sep);
+}
